Get My Usages Endpoint (#8)
Self-explanatory. Currently paginated and sortable by time started,
cost, and machine name. Threw machine name in last minute, so if it
doesn't work properly just let me know and I'll patch it. Was gonna do a
sort by semester, but realistically you can just do a time_started sort
and that should be the same for 99% of cases.