Machine Fail Form (#9)
Created a machine failure form that can be accessed in a user's my
forge, so status page can redirect here later. It is connected to the
machine failure API endpoint. Right now it doesn't filter by which
machines have an active usage, but that should get added at some point.

Also fixed authentication time logout so it isn't as short and so it's a
const variable and not a magic number.

**Note:** this branch has pulls from Machine Usage Form #6, so that PR
should probably be merged first.

---------

Co-authored-by: Thomas Byrne <tbyrne423@yahoo.com>